pyproject.toml
[tool.poetry]
name = "summarisr"
version = "0.1.0"
description = "This is a simple HTTP REST service that stores english texts and returns a summary of them."
authors = ["Mikio Oba <hello@mikio.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
sumy = "^0.10.0"
numpy = "^1.23.0"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
python-multipart = "^0.0.5"
pytest-cov = "^3.0.0"


[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.18.3"
pytest-dotenv = "^0.5.2"
pytest-docker-compose = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
